Refactor Command.cs into its own namespace and files
