+++ S2VX.Game/Story/Command/CommandType.cs
@@ -1,0 +1,20 @@
+﻿namespace S2VX.Game.Story.Command {
+    public enum CommandType {
+        None,
+        ApproachesDistance,
+        ApproachesThickness,
+        BackgroundColor,
+        CameraMove,
+        CameraRotate,
+        CameraScale,
+        GridAlpha,
+        GridColor,
+        GridThickness,
+        NotesAlpha,
+        NotesColor,
+        NotesFadeInTime,
+        NotesFadeOutTime,
+        NotesShowTime,
+        TimingChange,
+    }
+}
