+++ S2VX.Game/Story/Command/GridCommand.cs
@@ -1,0 +1,5 @@
+﻿namespace S2VX.Game.Story.Command {
+    public abstract class GridCommand : S2VXCommand {
+        public Grid Grid { get; } = new Grid();
+    }
+}
