+++ S2VX.Game/Story/Command/GridThicknessCommand.cs
@@ -1,0 +1,22 @@
+﻿using osu.Framework.Utils;
+using System.Globalization;
+
+namespace S2VX.Game.Story.Command {
+    public class GridThicknessCommand : S2VXCommand {
+        public override CommandType Type { get; set; } = CommandType.GridThickness;
+        public float StartValue { get; set; } = 0.005f;
+        public float EndValue { get; set; } = 0.005f;
+        public override void Apply(double time, S2VXStory story) {
+            var thickness = Interpolation.ValueAt(time, StartValue, EndValue, StartTime, EndTime, Easing);
+            story.Grid.Thickness = thickness;
+        }
+        protected override string ToValues() => $"{StartValue}|{EndValue}";
+        public static GridThicknessCommand FromString(string[] split) {
+            var command = new GridThicknessCommand() {
+                StartValue = float.Parse(split[4], CultureInfo.InvariantCulture),
+                EndValue = float.Parse(split[5], CultureInfo.InvariantCulture),
+            };
+            return command;
+        }
+    }
+}
