+++ S2VX.Game/Story/S2VXStory.cs
@@ -4,6 +4,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osuTK.Graphics;
+using S2VX.Game.Story.Command;
 using System.Collections.Generic;
 using System.IO;
 
@@ -22,9 +23,9 @@ namespace S2VX.Game.Story {
         public Notes Notes { get; } = new Notes();
         public Approaches Approaches { get; } = new Approaches();
 
-        public List<Command> Commands { get; private set; } = new List<Command>();
+        public List<S2VXCommand> Commands { get; private set; } = new List<S2VXCommand>();
         private int NextActive { get; set; }
-        private HashSet<Command> Actives { get; set; } = new HashSet<Command>();
+        private HashSet<S2VXCommand> Actives { get; set; } = new HashSet<S2VXCommand>();
 
         private static JsonConverter[] Converters { get; } = { new Vector2Converter(), new NoteConverter() };
 
@@ -46,7 +47,7 @@ namespace S2VX.Game.Story {
             Actives.Clear();
         }
 
-        public void AddCommand(Command command) {
+        public void AddCommand(S2VXCommand command) {
             Commands.Add(command);
             Commands.Sort();
             ClearActives();
@@ -74,7 +75,7 @@ namespace S2VX.Game.Story {
             var story = JObject.Parse(text);
             var serializedCommands = JsonConvert.DeserializeObject<List<JObject>>(story[nameof(Commands)].ToString());
             foreach (var serializedCommand in serializedCommands) {
-                var command = Command.FromJson(serializedCommand);
+                var command = S2VXCommand.FromJson(serializedCommand);
                 Commands.Add(command);
             }
             Commands.Sort();
@@ -105,7 +106,7 @@ namespace S2VX.Game.Story {
                 Actives.Add(Commands[NextActive++]);
             }
 
-            var newActives = new HashSet<Command>();
+            var newActives = new HashSet<S2VXCommand>();
             foreach (var active in Actives) {
                 // Run active commands
                 active.Apply(Time.Current, this);
